+++ src/Validator/Constraints/Date/LaterThanTodayValidator.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * (c) Meritoo.pl, http://www.meritoo.pl
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Meritoo\CommonBundle\Validator\Constraints\Date;
+
+/**
+ * Validator of date that should be later than today
+ *
+ * @author    Meritoo <github@meritoo.pl>
+ * @copyright Meritoo <http://www.meritoo.pl>
+ *
+ * @Annotation
+ */
+class LaterThanTodayValidator extends BaseThanTodayValidator
+{
+}